# server/swagger.yaml
openapi: 3.0.0
info:
  title: ModernDoc API
  description: API endpoint for ModernDoc
servers:
  - url: http://localhost:8080
    description: ModernDoc API
...
paths:
...
  /field/create:
    post:
      tags:
        - Field
      summary: Create a new field
      operationId: createField
      requestBody:
        description: Field object that needs to be added
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  description: Field's name
                officialName:
                  type: string
                  description: Field's official name
                description:
                  type: string
                  description: Field's description
                type:
                  type: string
                  description: text,number,date,singleSelect,multipleSelect
      responses:
        '200':
          description: Field created
        '500':
          description: Field creation failed
...
  /field/edit-official-name:
    put:
      tags:
        - Field
      summary: Edit field's official name
      operationId: editFieldOfficialName
      requestBody:
        description: Field object that needs to be edited
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  type: string
                  description: Field's id
                officialName:
                  type: string
                  description: Field's official name
      responses:
        '200':
          description: Field edited
        '500':
          description: Field edit failed
...
  /folder/get-by-id/{id}:
    get:
      tags:
        - Folder
      summary: Get folder by id
      operationId: getFolderById
      parameters:
        - name: id
          in: path
          description: Folder's id
          required: true
          schema:
            type: string
        - name: user-id
          in: headers
          description: User's id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Folder retrieved
        '400':
          description: Folder retrieval failed
...